init and confirm DataStores directly on DLSM
Rather than forwarding these calls from the QueryManager to the DataLayrServiceManager, instead send them directly to the DataLayrServiceManager. This change requires reducing a bunch of constants in the DataLayrSignatureChecker.sol by 32, since the calldata is now 32 bytes shorter.
All tests are passing; this also decreases the cost of initing or confirming a DataStore by ~1k gas