create abstract (for now) RegistrationManagerBase
In an effort to build base contracts for general middlewares, this takes some of the logic contained in DataLayrVoteWeigher and moves it into a simple, general-purpose contract. DataLayrVoteWeigher now inherits from this base contract, similar to how it inherits from VoteWeigherBase.
Currently 'RegistrationManagerBase' is abstract, since it does not implement a couple functions, but I've left a TODO for making this non-abstract