move 'slashedStatus' storage + updates from 'InvestmentManager' to 'Slasher'
This **really** seems like the proper place for this storage to live. The change means that the slasher no longer just passes calls onto `investmentManager.slashOperator`, and that the slasher (rather than the InvestmentManager) defines functionality for both `hasBeenSlashed` and `resetSlashedStatus`.
Most changes in this commit are minor and just due to knock-on effects of the primary change.