get rid of `serviceFactories` mapping in Slasher contract
-remove related functions `addserviceFactories`, `removeserviceFactories`, `canSlash(address toBeSlashed, IServiceFactory serviceFactory, IRepository repository, IRegistry registry)`, `slashOperator(address toBeSlashed, IServiceFactory serviceFactory, IRepository repository, IRegistry registry)`
-modify InvestmentManager, EigenLayrDelegation, and testing contracts to work with this change
I still think the existing pattern will need more work, but AFAICT this (removed) pattern is not one we'll want to launch with -- it'd be good to be able to add it back later, but I think it's unnecessary to launch with it.
All tests are still passing with this change.