combine 'Repository' and 'RepositoryStorage' contracts
Repository was not created to be upgradeable -- rather, it is Initializable so that it can be deployed by a factory and hopefully auto-verified in block explorers like Etherscan (since all copies will have the same constructor params). Since we are using the pattern of a separate storage contract for upgradeability, I think this merge makes sense and slightly improves code readability.