naming change: attempt at clarity in 2-step slashing
Previously we were using 'slashed/slashing/slash' *both* for the 1st step (funds frozen) *and* for the 2nd step (actually 'slashing' funds, a.k.a. the operator or staker actually losing their shares).
This commit changes the language around the 1st step to instead be 'frozen/freezing/freeze'.
Some comments may still be incorrect, but hopefully this helps with clarity.