move struct definitions to interface
also mark 'totalStakeHistory' array as internal, since 'getTotalStakeFromIndex' is a getter function for it. Solidity won't let you override a function that returns a struct with a dynamic array -- we'd have to make its return type (uint32, uint32, uint96, uint96), which is just kinda ugly & annoying to deal with vs. this solution