make investmentManager immutable in InvestmentStrategyBase
this breaks a ton of stuff due to percolation, but I think it'll be worth it in terms of gas savings, and makes sense since the investmentManager is going to be an upgradable proxy. I'll go through and fix the new issues in another commit.
This commit also adds some better revert messages (up to our new standard)