purposefully committing failing test
The `testFraudproofQueuedWithdrawal` test is still failing and definitely a bit of a mess in this state with lots of copy-pasted code, but I believe it makes a bug clear -- I've pointed out how the order of variables used in the `stakeWithdrawalVerification` function is *different* from that used in the `computeDataStoreHash` function, and believe this to be the root cause of the current failure.
Sorting out and fixing this discrepancy is a next step.