add solhint support
modify 'max-line-length' to be a *warning* instead of an error, and adjust to a more sane length (someone else can pare this down a little if they think '200' is too large).
split a couple of the existing longest lines into multiple lines, so they fall under the new line-length limit.