fix flaky test
This test was failing sometimes for me. Realized the failing calls were when it made a call to 'address(2)' (which is a precompile address) -- it appears that  this address passes the assume statement
```
        // assume that the fuzzed address is not already a contract!
        uint256 size;
        assembly {
            size := extcodesize(nonexistentStrategy)
        }
        cheats.assume(size == 0);
```
but does not revert when called by the investmentManager inside of its `investmentManager`.
I suspect this is because address(2) is the precompile for SHA256 and thus returns the correct amount of data for the call made by the investmentManager here:
```
        // deposit the assets into the specified strategy and get the equivalent amount of
        // shares in that strategy
        shares = strategy.deposit(token, amount);
```