make BLSRegistryWithBomb upgradeable
This is a pretty big change.
Due to inheritance patterns + what I think makes the most sense, VoteWeigherBase and RegistryBase are now also implemented as upgradeable, abstract contracts, with internal `_initialize` functions.
Otherwise this is primarily just making the  BLSRegistry and BLSRegistryWithBomb work with the same kind of pattern we're using elsewhere.