potentially correct fallback routine
I've done my best to comment this. Basically, rather than reverting (using require statements), we keep track of whether the `insertAfter` input is incorrect using a boolean named `runFallbackRoutine`. If `runFallbackRoutine` remains 'false', then we proceed as we did before. If it ever gets marked as 'true' (i.e. when we would previously have reverted!), then we run `_getCorrectValueForInsertAfter` to search for the correct value, update the value of `insertAfter`, and run the logic again.