implement most of the separate pausing switches
I got a 'stack-too-deep' error in trying to apply `onlyWhenNotPaused(PAUSED_WITHDRAWALS)` instead of `whenNotPaused` to the `queueWithdrawal` function of the InvestmentManager. Couldn't figure out how to resolve it so far; seems to need more work.
Also had to modify a (unfinished!) test to work a bit better with the new design. Tests will need additional updates to actually reach good coverage.