split up the deployment of EigenLayr and DataLayr
This is a preliminary step to splitting up the test files. Likely there is more storage that can be moved from the EigenLayrDeployer to the DataLayrDeployer. This was just what I got to work for now.
Note that the `TestHelper` still has to inherit from the `DataLayrDeployer` in this design -- there's more work to do on this.