Add 'feeSetter' role to DLSM
This new role is now the only address that can call `setFeePerBytePerTime`. The address can be adjusted by the DLSM owner through the new `setFeeSetter` function.
Adding the role like this made more sense to me than doing a 'quorumMaster'-type role as had been previously suggested. I am open to other options, but this is a minimal change that seems to work with what we had discussed.