Initial Accounting Fix
This set of changes significantly simplifies the accounting in EigenPods.
Key changes are to the EigenPod and InvestmentManager contracts. EigenPodManager also has small changes due to eliminating the need for a couple functions. This commit includes documentation changes, but the docs will need more subsequent improvement + clarification.

Rather than charging "penalties" when podOwners are "overcommitted", instead we now just pull their shares from EigenLayer until the validator has withdrawn. This solution is far friendlier to stakers, while being what seems to be an acceptable tradeoff for middlewares.
Specifically, all penalty-related logic has been removed from the `verifyOvercommittedStake` function, but it continues to decrement the user's beaconChainETH shares through a call to `eigenPodManager.recordOvercommittedBeaconChainETH`.

There are now 4 cases for a withdrawing validator:
1) happy path (not proven overcommitted, not overcommitted when withdrawing) -- this works as before (except without penalties) -- REQUIRED_BALANCE_WEI is held for normal withdrawal path, everything in excess becomes instantly withdrawable
2) overcommitted, but not previously proven to be overcommitted -- the full amount is held for the normal withdrawal path, and all excess shares are removed from EigenLayer
3) previously proven overcommitted, but no longer overcommitted -- just like (1) above, except the podOwner's beaconChainShares are increased by REQUIRED_BALANCE_WEI (to counterbalance the decrease when they were proven to be overcommitted)
4) previously proven overcommitted, still overcommitted -- the full amount is held for the normal withdrawal path, and the podOwner is credited with enough beaconChainETH shares in EigenLayer to complete the normal withdrawal process

Note as well the change to InvestmentManager -- the introduction of the `beaconChainETHWithdrawalDebt` mapping helps deal with a potential accounting issue when attempting to decrement a staker's beaconChainETH shares.