reduce overly pessimistic blockNumber checks and introduce VERIFY_OVERCOMMITTED_WINDOW_BLOCKS variable
We actually only need to check against `mostRecentWithdrawalBlockNumber` in the `verifyCorrectWithdrawalCredentials` function.
`verifyAndProcessWithdrawal` does not need this check, since the validator must have already been proven using `verifyCorrectWithdrawalCredentials` and a validator can only be fully withdrawn from one time.
`verifyOvercommittedStake` is intended to verify only if a validator is/was very recently overcommitted, so this commit add the `VERIFY_OVERCOMMITTED_WINDOW_BLOCKS` constant and checks against that instead.